[🚚move] 기존 tickets 폴더명 변경 (samples)
- 기존 샘플 파일들 samples 디렉토리로 변경
