+++ src/components/samples/TicketItem.tsx
