+++ src/index.tsx
@@ -1,7 +1,7 @@
 /** pages */
 export { GlobalNavigation } from '@components/GlobalNavigation';
 export { GlobalHeader } from '@components/header/GlobalHeader.tsx';
-export { SubHeader } from '@components/header/SubHeader';
+export { SubHeader } from '@components/header/subHeader';
 export { Layout } from '@components/layout/Layout';
 export { Home } from './pages/Home';
 export { Login } from './pages/Login';
