+++ src/pages/Center.tsx
@@ -1,0 +1,3 @@
+export const Center = () => {
+  return null;
+};
